Fix building for debian 13 armv7l
In Debian 13 armv7l, time_t and suseconds_t are "long long int" instead
of "long unsigned int" as seen on x86_64 and aarch64. Fix it by casting
to "long unsigned int".

Related: OS#6828
(cherry picked from commit 8baa7cec367a43cb1f22016e3e2192abdcd9fe8c)

+++ src/osmo-bsc/bts_ctrl.c
@@ -130,7 +130,7 @@ static int get_bts_loc(struct ctrl_cmd *cmd, void *data)
 
 	curloc = llist_first_entry(&bts->loc_list, struct bts_location, list);
 
-	cmd->reply = talloc_asprintf(cmd, "%lu,%s,%f,%f,%f", curloc->tstamp,
+	cmd->reply = talloc_asprintf(cmd, "%lu,%s,%f,%f,%f", (unsigned int long) curloc->tstamp,
 			get_value_string(bts_loc_fix_names, curloc->valid), curloc->lat, curloc->lon, curloc->height);
 	if (!cmd->reply) {
 		cmd->reply = "OOM";
