+++ tests/acc/acc_test.c
@@ -36,7 +36,7 @@ static void clock_debug(char* str)
 	struct timeval tv;
 	osmo_gettimeofday(&tv, NULL);
 	fprintf(stderr, "sys={%lu.%06lu}: %s\n",
-		tv.tv_sec, tv.tv_usec, str);
+		(unsigned int long) tv.tv_sec, (unsigned int long) tv.tv_usec, str);
 }
 
 #define bts_init(net) _bts_init(net, __func__)
