+++ tests/paging/paging_test.c
@@ -42,7 +42,7 @@ static void clock_debug(char *str)
 	struct timeval tv;
 	osmo_gettimeofday(&tv, NULL);
 	fprintf(stderr, "sys={%lu.%06lu}: %s\n",
-		tv.tv_sec, tv.tv_usec, str);
+		(unsigned int long) tv.tv_sec, (unsigned int long) tv.tv_usec, str);
 }
